+++ example/cli/src/app/app.module.ts
@@ -20,6 +20,10 @@ import { DragComponent } from './drag/drag.component';
 import { VirtualscrollComponent } from './virtualscroll/virtualscroll.component';
 import { ApiComponent } from './api/api.component';
 import { ActionsComponent } from './actions/actions.component';
+import { ScrollContainerComponent } from './scrollcontainer/scrollcontainer.component';
+import { ContextmenuComponent } from './contextmenu/contextmenu.component';
+import { DragOverStylingComponent } from './dragover-styling/dragover-styling.component';
+import { DragOverStylingFullTreeComponent } from './dragover-styling/dragover-styling-full-tree.component';
 
 @NgModule({
   declarations: [
@@ -35,18 +39,22 @@ import { ActionsComponent } from './actions/actions.component';
     SaveRestoreComponent,
     CheckboxesComponent,
     DragComponent,
+    DragOverStylingComponent,
+    DragOverStylingFullTreeComponent,
     VirtualscrollComponent,
     ApiComponent,
-    ActionsComponent
+    ActionsComponent,
+    ScrollContainerComponent,
+    ContextmenuComponent
   ],
   imports: [
     BrowserModule,
     FormsModule,
-    TreeModule,
+    TreeModule.forRoot(),
     CommonModule,
     AppRoutingModule
   ],
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}