+++ example/cli/src/app/checkboxes/checkboxes.component.ts
@@ -6,6 +6,7 @@ import { Component } from '@angular/core';
   template: `
     <h3>tri-state checkboxes</h3>
     <tree-root
+      id="tree1"
       [nodes]="nodes"
       [options]="options">
     </tree-root>
@@ -13,8 +14,15 @@ import { Component } from '@angular/core';
     Switch expander and checkbox with CSS 'order' attribute:</h3>
     <tree-root
       class="reverse"
+      id="tree2"
       [nodes]="nodes"
       [options]="options">
+    </tree-root>
+    <h3>Disable tri-state checkboxes</h3>
+    <tree-root
+      id="tree3"
+      [nodes]="nodes"
+      [options]="optionsDisabled">
     </tree-root>
   `,
   styles: [
@@ -23,7 +31,7 @@ import { Component } from '@angular/core';
 export class CheckboxesComponent {
   nodes = [
     {
-      name: 'root1'
+      name: 'root1',
     },
     {
       name: 'root2',
@@ -34,10 +42,36 @@ export class CheckboxesComponent {
           { name: 'grandchild2' }
         ] }
       ]
+    },
+    {
+      name: 'asyncroot',
+      hasChildren: true
     }
   ];
 
   options: ITreeOptions = {
-    useCheckbox: true
+    useCheckbox: true,
+    getChildren: this.getChildren.bind(this)
   };
+
+  optionsDisabled: ITreeOptions = {
+    useCheckbox: true,
+    getChildren: this.getChildren.bind(this),
+    useTriState: false
+  };
+
+  getChildren(node: any) {
+    const newNodes = [
+      {
+        name: 'child1'
+      }, {
+        name: 'child2'
+      }
+    ];
+
+    return new Promise((resolve, reject) => {
+      setTimeout(() => resolve(newNodes), 1000);
+    });
+  }
+
 }