+++ lib/components/tree-viewport.component.ts
@@ -1,15 +1,23 @@
 import {
-  Component, ElementRef, ViewEncapsulation, HostListener, AfterViewInit, OnInit, OnDestroy
+  Component,
+  ElementRef,
+  ViewEncapsulation,
+  AfterViewInit,
+  OnInit,
+  OnDestroy,
+  NgZone
 } from '@angular/core';
 import { TreeVirtualScroll } from '../models/tree-virtual-scroll.model';
 import { TREE_EVENTS } from '../constants/events';
+import { Cancelable } from 'lodash';
+import throttle from 'lodash/throttle';
 
 @Component({
   selector: 'tree-viewport',
   styles: [],
   providers: [TreeVirtualScroll],
   template: `
-    <ng-container *mobxAutorun="{dontDetach: true}">
+    <ng-container *treeMobxAutorun="{ dontDetach: true }">
       <div [style.height]="getTotalHeight()">
         <ng-content></ng-content>
       </div>
@@ -17,9 +25,17 @@ import { TREE_EVENTS } from '../constants/events';
   `
 })
 export class TreeViewportComponent implements AfterViewInit, OnInit, OnDestroy {
+  setViewport = throttle(() => {
+    this.virtualScroll.setViewport(this.elementRef.nativeElement);
+  }, 17);
+  private readonly scrollEventHandler: ($event: Event) => void;
+
   constructor(
     private elementRef: ElementRef,
-    public virtualScroll: TreeVirtualScroll) {
+    private ngZone: NgZone,
+    public virtualScroll: TreeVirtualScroll
+  ) {
+    this.scrollEventHandler = this.setViewport.bind(this);
   }
 
   ngOnInit() {
@@ -31,22 +47,23 @@ export class TreeViewportComponent implements AfterViewInit, OnInit, OnDestroy {
       this.setViewport();
       this.virtualScroll.fireEvent({ eventName: TREE_EVENTS.initialized });
     });
+    let el: HTMLElement = this.elementRef.nativeElement;
+    this.ngZone.runOutsideAngular(() => {
+      el.addEventListener('scroll', this.scrollEventHandler);
+    });
   }
 
   ngOnDestroy() {
     this.virtualScroll.clear();
+    let el: HTMLElement = this.elementRef.nativeElement;
+    el.removeEventListener('scroll', this.scrollEventHandler);
   }
 
   getTotalHeight() {
-    return this.virtualScroll.isEnabled() && this.virtualScroll.totalHeight + 'px' || 'auto';
-  }
-
-  @HostListener('scroll', ['$event'])
-  onScroll() {
-    this.setViewport();
-  }
-
-  setViewport() {
-    this.virtualScroll.setViewport(this.elementRef.nativeElement);
+    return (
+      (this.virtualScroll.isEnabled() &&
+        this.virtualScroll.totalHeight + 'px') ||
+      'auto'
+    );
   }
 }