+++ lib/defs/api.ts
@@ -2,7 +2,7 @@
  * Welcome to ng2tree
  */
 export type IDType = string | number;
-export type IDTypeDictionary = { [id: string]: boolean } | { [id: number]: boolean };
+export type IDTypeDictionary = { [id: string]: boolean, [id: number]: boolean };
 
 /**
  * See ITreeNode for documentation
@@ -24,8 +24,10 @@ export interface IAllowDragFn {
 
 export interface ITreeState {
   expandedNodeIds?: IDTypeDictionary;
+  selectedNodeIds?: IDTypeDictionary;
   activeNodeIds?: IDTypeDictionary;
   hiddenNodeIds?: IDTypeDictionary;
+  selectedLeafNodeIds?: IDTypeDictionary;
   focusedNodeId?: IDType;
 }
 
@@ -140,6 +142,23 @@ export interface ITreeOptions {
     ```
    */
    allowDrop?: boolean | IAllowDropFn;
+    /**
+    Boolean flag to allow adding and removing is-dragging-over and is-dragging-over-disabled classes.
+
+    If set to false it will not add the above mentioned classes and you should handle the styling yourself with css and in
+    the actionMapping -> mouse -> dragEnter, dragLeave
+
+    * **Default Value: true**
+
+    example:
+    ```
+    * options = {
+    *   allowDrop: true,
+    *   allowDragoverStyling: false
+    * }
+    ```
+    */
+   allowDragoverStyling?: boolean;
    /**
    * Specify padding per node (integer).
     Each node will have padding-left value of level * levelPadding, instead of using the default padding for children.
@@ -218,11 +237,11 @@ export interface ITreeOptions {
     */
    animateAcceleration?: number;
    /**
-    * Whether to scroll to the node to make it visible when it is selected / activated.
+    * Whether to scroll to the node to make it visible when it is activated.
 
     * **Default Value: true**
     */
-    scrollOnSelect?: boolean;
+   scrollOnActivate?: boolean;
    /**
     * Function to clone a node.
     * Receives a TreeNode object, and returns a node object (only the data).
@@ -256,6 +275,16 @@ export interface ITreeOptions {
      * Whether to display a checkbox next to the node or not
      */
     useCheckbox?: boolean;
+    /**
+     * Whether to use master checkboxes mechanism if the useCheckbox is set to true
+     */
+    useTriState?: boolean;
+    /**
+     * The HTML element that is the scroll container for the tree.
+     * The default behaviour is to wrap the tree with a container that has overflow: hidden,
+     * and then the scrolling container is the viewport inside the tree component
+     */
+    scrollContainer?: HTMLElement;
  }
 
 export interface ITreeNode {
@@ -453,7 +482,7 @@ export interface ITreeModel {
    */
   isFocused: boolean;
   /**
-   * @returns Current active (selected) nodes
+   * @returns Current active nodes
    */
   activeNodes: ITreeNode[];
   /**
@@ -463,7 +492,7 @@ export interface ITreeModel {
 
   // helpers
   /**
-   * @returns Current active (selected) node. If multiple nodes are active - returns the first one.
+   * @returns Current active node. If multiple nodes are active - returns the first one.
    */
   getActiveNode(): ITreeNode;
   /**
@@ -511,6 +540,10 @@ export interface ITreeModel {
    * @returns   First node that matches the predicate, if found - null otherwise
    */
   getNodeBy(predicate: any, startNode?: ITreeNode): ITreeNode;
+  /**
+   * get tree state
+   */
+  getState(): ITreeState;
 
   // actions
   /**
@@ -551,9 +584,10 @@ export interface ITreeModel {
    * moves a node from one location in the tree to another
    * @param node describes which node needs to be moved
    * @param to describes where to move the node to.
+   * @param from describes where to move the node from.
    * Contains a 'parent' node, an 'index', and a 'dropOnNode' - to distinguish between dropping between nodes or on the node
    */
-  moveNode(node: ITreeNode, to: {parent: ITreeNode, index: number, dropOnNode: boolean});
+  moveNode(node: ITreeNode, to: {parent: ITreeNode, index: number, dropOnNode: boolean}, from: {parent: ITreeNode, index: number});
   /**
    * Invokes a method for every node of the tree - depth first
    * @param fn  a function that receives the node
@@ -567,10 +601,6 @@ export interface ITreeModel {
    * collapse all nodes
    */
   collapseAll();
-  /**
-   * get tree state
-   */
-  getState(): ITreeState;
   /**
    * set tree state
    */