+++ lib/directives/tree-drag.directive.ts
@@ -1,4 +1,4 @@
-import { Directive, Input, HostListener, Renderer, ElementRef, DoCheck } from '@angular/core';
+import { AfterViewInit, Directive, DoCheck, ElementRef, HostListener, Input, NgZone, OnDestroy, Renderer2 } from '@angular/core';
 import { TreeDraggedElement } from '../models/tree-dragged-element.model';
 
 const DRAG_OVER_CLASS = 'is-dragging-over';
@@ -6,15 +6,29 @@ const DRAG_OVER_CLASS = 'is-dragging-over';
 @Directive({
   selector: '[treeDrag]'
 })
-export class TreeDragDirective implements DoCheck {
+export class TreeDragDirective implements AfterViewInit, DoCheck, OnDestroy {
   @Input('treeDrag') draggedElement;
   @Input() treeDragEnabled;
+  private readonly dragEventHandler: (ev: DragEvent) => void;
 
-  constructor(private el: ElementRef, private renderer: Renderer, private treeDraggedElement: TreeDraggedElement) {
+  constructor(private el: ElementRef, private renderer: Renderer2, private treeDraggedElement: TreeDraggedElement, private ngZone: NgZone) {
+    this.dragEventHandler = this.onDrag.bind(this);
+  }
+
+  ngAfterViewInit() {
+    let el: HTMLElement = this.el.nativeElement;
+    this.ngZone.runOutsideAngular(() => {
+      el.addEventListener('drag', this.dragEventHandler);
+    });
   }
 
   ngDoCheck() {
-    this.renderer.setElementAttribute(this.el.nativeElement, 'draggable', this.treeDragEnabled ? 'true' : 'false');
+    this.renderer.setAttribute(this.el.nativeElement, 'draggable', this.treeDragEnabled ? 'true' : 'false');
+  }
+
+  ngOnDestroy() {
+    let el: HTMLElement = this.el.nativeElement;
+    el.removeEventListener('drag', this.dragEventHandler);
   }
 
   @HostListener('dragstart', ['$event']) onDragStart(ev) {
@@ -22,13 +36,13 @@ export class TreeDragDirective implements DoCheck {
     ev.dataTransfer.setData('text', ev.target.id);
     this.treeDraggedElement.set(this.draggedElement);
     if (this.draggedElement.mouseAction) {
-        this.draggedElement.mouseAction('dragStart', ev);
+      this.draggedElement.mouseAction('dragStart', ev);
     }
   }
 
-  @HostListener('drag', ['$event']) onDrag(ev) {
+  onDrag(ev) {
     if (this.draggedElement.mouseAction) {
-        this.draggedElement.mouseAction('drag', ev);
+      this.draggedElement.mouseAction('drag', ev);
     }
   }