+++ lib/directives/tree-drop.directive.ts
@@ -1,4 +1,15 @@
-import { Directive, Output, Input, EventEmitter, HostListener, Renderer, ElementRef } from '@angular/core';
+import {
+  AfterViewInit,
+  Directive,
+  ElementRef,
+  EventEmitter,
+  HostListener,
+  Input,
+  NgZone,
+  OnDestroy,
+  Output,
+  Renderer2
+} from '@angular/core';
 import { TreeDraggedElement } from '../models/tree-dragged-element.model';
 
 const DRAG_OVER_CLASS = 'is-dragging-over';
@@ -7,47 +18,86 @@ const DRAG_DISABLED_CLASS = 'is-dragging-over-disabled';
 @Directive({
   selector: '[treeDrop]'
 })
-export class TreeDropDirective {
+export class TreeDropDirective implements AfterViewInit, OnDestroy {
+  @Input() allowDragoverStyling = true;
   @Output('treeDrop') onDropCallback = new EventEmitter();
   @Output('treeDropDragOver') onDragOverCallback = new EventEmitter();
   @Output('treeDropDragLeave') onDragLeaveCallback = new EventEmitter();
   @Output('treeDropDragEnter') onDragEnterCallback = new EventEmitter();
+  private readonly dragOverEventHandler: (ev: DragEvent) => void;
+  private readonly dragEnterEventHandler: (ev: DragEvent) => void;
+  private readonly dragLeaveEventHandler: (ev: DragEvent) => void;
 
   private _allowDrop = (element, $event) => true;
+
   @Input() set treeAllowDrop(allowDrop) {
     if (allowDrop instanceof Function) {
       this._allowDrop = allowDrop;
     }
     else this._allowDrop = (element, $event) => allowDrop;
   }
+
   allowDrop($event) {
     return this._allowDrop(this.treeDraggedElement.get(), $event);
   }
 
-  constructor(private el: ElementRef, private renderer: Renderer, private treeDraggedElement: TreeDraggedElement) {
+  constructor(private el: ElementRef, private renderer: Renderer2, private treeDraggedElement: TreeDraggedElement, private ngZone: NgZone) {
+    this.dragOverEventHandler = this.onDragOver.bind(this);
+    this.dragEnterEventHandler = this.onDragEnter.bind(this);
+    this.dragLeaveEventHandler = this.onDragLeave.bind(this);
   }
 
-  @HostListener('dragover', ['$event']) onDragOver($event) {
-    if (!this.allowDrop($event)) return this.addDisabledClass();
+  ngAfterViewInit() {
+    let el: HTMLElement = this.el.nativeElement;
+    this.ngZone.runOutsideAngular(() => {
+      el.addEventListener('dragover', this.dragOverEventHandler);
+      el.addEventListener('dragenter', this.dragEnterEventHandler);
+      el.addEventListener('dragleave', this.dragLeaveEventHandler);
+    });
+  }
+
+  ngOnDestroy() {
+    let el: HTMLElement = this.el.nativeElement;
+    el.removeEventListener('dragover', this.dragOverEventHandler);
+    el.removeEventListener('dragenter', this.dragEnterEventHandler);
+    el.removeEventListener('dragleave', this.dragLeaveEventHandler);
+  }
+
+  onDragOver($event) {
+    if (!this.allowDrop($event)) {
+      if (this.allowDragoverStyling) {
+        return this.addDisabledClass();
+      }
+      return;
+    }
 
     this.onDragOverCallback.emit({event: $event, element: this.treeDraggedElement.get()});
 
     $event.preventDefault();
-    this.addClass();
+    if (this.allowDragoverStyling) {
+      this.addClass();
+    }
   }
 
-  @HostListener('dragenter', ['$event']) onDragEnter($event) {
+  onDragEnter($event) {
     if (!this.allowDrop($event)) return;
 
+    $event.preventDefault();
     this.onDragEnterCallback.emit({event: $event, element: this.treeDraggedElement.get()});
   }
 
-  @HostListener('dragleave', ['$event']) onDragLeave($event) {
-    if (!this.allowDrop($event)) return this.removeDisabledClass();
-
+  onDragLeave($event) {
+    if (!this.allowDrop($event)) {
+      if (this.allowDragoverStyling) {
+        return this.removeDisabledClass();
+      }
+      return;
+    }
     this.onDragLeaveCallback.emit({event: $event, element: this.treeDraggedElement.get()});
 
-    this.removeClass();
+    if (this.allowDragoverStyling) {
+      this.removeClass();
+    }
   }
 
   @HostListener('drop', ['$event']) onDrop($event) {
@@ -55,23 +105,26 @@ export class TreeDropDirective {
 
     $event.preventDefault();
     this.onDropCallback.emit({event: $event, element: this.treeDraggedElement.get()});
-    this.removeClass();
+
+    if (this.allowDragoverStyling) {
+      this.removeClass();
+    }
     this.treeDraggedElement.set(null);
   }
 
   private addClass() {
-    this.renderer.setElementClass(this.el.nativeElement, DRAG_OVER_CLASS, true);
+    this.renderer.addClass(this.el.nativeElement, DRAG_OVER_CLASS);
   }
 
   private removeClass() {
-    this.renderer.setElementClass(this.el.nativeElement, DRAG_OVER_CLASS, false);
+    this.renderer.removeClass(this.el.nativeElement, DRAG_OVER_CLASS);
   }
 
   private addDisabledClass() {
-    this.renderer.setElementClass(this.el.nativeElement, DRAG_DISABLED_CLASS, true);
+    this.renderer.addClass(this.el.nativeElement, DRAG_DISABLED_CLASS);
   }
 
   private removeDisabledClass() {
-    this.renderer.setElementClass(this.el.nativeElement, DRAG_DISABLED_CLASS, false);
+    this.renderer.removeClass(this.el.nativeElement, DRAG_DISABLED_CLASS);
   }
 }